+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "yaml-plus-json" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+## [0.1.0] 2019-02-8
+
+### Added
+
+* Right click `.json` files to convert to `.yml`.
+* Right click `.yml` or `.yaml` files to convert to json.